moved all view specific logic to the separate namespace
