+++ src/CoreCraft.SourceGeneration/Generators/ModelGenerator.cs
@@ -17,6 +17,7 @@ internal static class ModelGenerator
         {
             code.WriteLine("using CoreCraft;");
             code.WriteLine("using CoreCraft.Core;");
+            code.WriteLine("using CoreCraft.Views;");
             code.WriteLine("using CoreCraft.ChangesTracking;");
             code.WriteLine("using CoreCraft.Persistence;");
             code.WriteLine("using CoreCraft.Persistence.History;");
