+++ src/CoreCraft/Subscription/Builders/RelationSubscriptionBuilder.cs
@@ -1,4 +1,5 @@
 ﻿using CoreCraft.ChangesTracking;
+using CoreCraft.Views;
 
 namespace CoreCraft.Subscription.Builders;
