+++ src/CoreCraft/Subscription/CollectionSubscription.cs
@@ -1,4 +1,5 @@
 ﻿using CoreCraft.ChangesTracking;
+using CoreCraft.Views;
 
 namespace CoreCraft.Subscription;
