+++ tests/CoreCraft.Generators.Tests/VerifiedFiles/ApplicationModelGeneratorTests.OneModelFileSnapshotTest_file=AllFeatures.model.json#AllFeatures.g.verified.cs
@@ -15,6 +15,7 @@ namespace compilation.AllFeatures
 {
     using CoreCraft;
     using CoreCraft.Core;
+    using CoreCraft.Views;
     using CoreCraft.ChangesTracking;
     using CoreCraft.Persistence;
     using CoreCraft.Persistence.History;
