+++ tests/CoreCraft.Generators.Tests/VerifiedFiles/ApplicationModelGeneratorTests.OneModelFileSnapshotTest_file=LoadManually.model.json#LoadManually.g.verified.cs
@@ -15,6 +15,7 @@ namespace compilation.LoadManually
 {
     using CoreCraft;
     using CoreCraft.Core;
+    using CoreCraft.Views;
     using CoreCraft.ChangesTracking;
     using CoreCraft.Persistence;
     using CoreCraft.Persistence.History;
@@ -700,3 +701,4 @@ namespace compilation.LoadManually.Entities
 
 
 }
+
