+++ tests/CoreCraft.Generators.Tests/VerifiedFiles/ApplicationModelGeneratorTests.TwoModelFilesSnapshotTest#AnotherModelFile.g.verified.cs
@@ -15,6 +15,7 @@ namespace compilation.AnotherModelFile
 {
     using CoreCraft;
     using CoreCraft.Core;
+    using CoreCraft.Views;
     using CoreCraft.ChangesTracking;
     using CoreCraft.Persistence;
     using CoreCraft.Persistence.History;
