+++ tests/CoreCraft.Tests/UndoRedoDomainModelTests.cs
@@ -8,7 +8,9 @@ namespace CoreCraft.Tests;
 
 public class UndoRedoDomainModelTests
 {
+#pragma warning disable CS8618 // Non-nullable field must contain a non-null value when exiting constructor. Consider adding the 'required' modifier or declaring as nullable.
     private UndoRedoDomainModel _model;
+#pragma warning restore CS8618 // Non-nullable field must contain a non-null value when exiting constructor. Consider adding the 'required' modifier or declaring as nullable.
 
     [SetUp]
     public void Setup()
